nf_net_ctrl: Add NFP_NET_CFG_MBOX_CMD_DCB_UPDATE define If the update field has NFP_NET_CFG_UPDATE_MBOX set for PF vNIC reconfig messages, look for the generic mailbox and process the mailbox message. The location of the generic mailbox might be set in a TLV. If a TLV is not found, assume the default.
The DCB(Data Center Bridging)feature use the mailbox to pass the
related reconfig update information. So Add a MBOX_CMD number for DCB
updates, if the value of mailbox command field is number of DCB updates,
that means the DCB config updates and we need to process the mailbox
message to get the according update info of DCB config.

Signed-off-by: Zizhong Zhu <zizhong.zhu@corigine.com>